Separate website stuff and parser
