+++ howlongtobeat_oracle/__init__.py
@@ -1,1 +1,2 @@
-__version__ = '0.1.0'
+# flake8: noqa
+from .main import HowLongToBeatOracle
